+++ src/services/index.ts
@@ -3,7 +3,9 @@ import {
   startTunnelRegistryServer,
 } from '../lib/tunnel/tunnel-registry-server.js';
 import * as afc from './ios/afc/index.js';
+import * as crashReports from './ios/crash-reports/index.js';
 import * as diagnostics from './ios/diagnostic-service/index.js';
+import * as houseArrest from './ios/house-arrest/index.js';
 import * as mobileImageMounter from './ios/mobile-image-mounter/index.js';
 import * as powerAssertion from './ios/power-assertion/index.js';
 import * as syslog from './ios/syslog-service/index.js';
@@ -11,7 +13,9 @@ import * as tunnel from './ios/tunnel-service/index.js';
 import * as webinspector from './ios/webinspector/index.js';
 
 export {
+  crashReports,
   diagnostics,
+  houseArrest,
   mobileImageMounter,
   powerAssertion,
   syslog,